+++ crates/aiken-lang/src/tipo/expr.rs
@@ -11,11 +11,11 @@ use super::{
 use crate::{
     ast::{
         self, Annotation, ArgName, AssignmentKind, AssignmentPattern, BinOp, Bls12_381Point,
-        ByteArrayFormatPreference, CallArg, Curve, Function, IfBranch, LogicalOpChainKind, Pattern,
-        RecordUpdateSpread, Span, TraceKind, TraceLevel, Tracing, TypedArg, TypedCallArg,
-        TypedClause, TypedIfBranch, TypedPattern, TypedRecordUpdateArg, TypedValidator, UnOp,
-        UntypedArg, UntypedAssignmentKind, UntypedClause, UntypedFunction, UntypedIfBranch,
-        UntypedPattern, UntypedRecordUpdateArg,
+        ByteArrayFormatPreference, CallArg, Curve, Function, IfBranch, LogicalOpChainKind,
+        Namespace, Pattern, RecordUpdateSpread, Span, TraceKind, TraceLevel, Tracing, TypedArg,
+        TypedCallArg, TypedClause, TypedIfBranch, TypedPattern, TypedRecordUpdateArg,
+        TypedValidator, UnOp, UntypedArg, UntypedAssignmentKind, UntypedClause, UntypedFunction,
+        UntypedIfBranch, UntypedPattern, UntypedRecordUpdateArg,
     },
     builtins::{from_default_function, BUILTIN},
     expr::{FnStyle, TypedExpr, UntypedExpr},
@@ -404,7 +404,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 module_alias,
                 label,
                 ..
-            } => (Some(module_alias), label),
+            } => (Some(Namespace::Module(module_alias.to_string())), label),
 
             TypedExpr::Var { name, .. } => (None, name),
 
@@ -413,7 +413,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
 
         Ok(self
             .environment
-            .get_value_constructor(module, name, location)?
+            .get_value_constructor(module.as_ref(), name, location)?
             .field_map())
     }
 
@@ -792,12 +792,12 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         args: Vec<UntypedRecordUpdateArg>,
         location: Span,
     ) -> Result<TypedExpr, Error> {
-        let (module, name): (Option<String>, String) = match self.infer(constructor.clone())? {
+        let (module, name): (Option<Namespace>, String) = match self.infer(constructor.clone())? {
             TypedExpr::ModuleSelect {
                 module_alias,
                 label,
                 ..
-            } => (Some(module_alias), label),
+            } => (Some(Namespace::Module(module_alias)), label),
 
             TypedExpr::Var { name, .. } => (None, name),
 
@@ -1068,6 +1068,69 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             return shortcircuit;
         }
 
+        // In case where we find an uppercase var name in a record access chain, we treat the type
+        // as a namespace and lookup the next constructor as if it were imported from the module's
+        // the type originally belong to.
+        match container {
+            UntypedExpr::Var {
+                name: ref type_name,
+                location: type_location,
+            } if TypeConstructor::might_be(type_name) => {
+                return self.infer_type_constructor_access(
+                    (type_name, type_location),
+                    (
+                        &label,
+                        access_location.map(|start, end| (start + type_name.len() + 1, end)),
+                    ),
+                );
+            }
+
+            UntypedExpr::FieldAccess {
+                location: type_location,
+                label: ref type_name,
+                container: ref type_container,
+            } if TypeConstructor::might_be(type_name) => {
+                if let UntypedExpr::Var {
+                    name: ref module_name,
+                    location: module_location,
+                } = type_container.as_ref()
+                {
+                    if TypeConstructor::might_be(module_name) {
+                        return Err(Error::InvalidFieldAccess {
+                            location: access_location,
+                        });
+                    }
+
+                    // Lookup the module using the declared name (which may have been rebind with
+                    // 'as'), to obtain its _full unambiguous name_.
+                    let (_, module) = self
+                        .environment
+                        .imported_modules
+                        .get(module_name)
+                        .ok_or_else(|| {
+                            self.environment
+                                .err_unknown_module(module_name.to_string(), *module_location)
+                        })?;
+
+                    return self.infer_inner_type_constructor_access(
+                        (module.name.as_str(), *module_location),
+                        (
+                            type_name,
+                            type_location.map(|start, end| (start + module_name.len() + 1, end)),
+                        ),
+                        (
+                            &label,
+                            access_location.map(|start, end| {
+                                (start + module_name.len() + type_name.len() + 2, end)
+                            }),
+                        ),
+                    );
+                }
+            }
+
+            _ => (),
+        };
+
         // Attempt to infer the container as a record access. If that fails, we may be shadowing the name
         // of an imported module, so attempt to infer the container as a module access.
         // TODO: Remove this cloning
@@ -1075,7 +1138,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             Ok(record_access) => Ok(record_access),
 
             Err(err) => match container {
-                UntypedExpr::Var { name, location } => {
+                UntypedExpr::Var { name, location } if !TypeConstructor::might_be(&name) => {
                     let module_access =
                         self.infer_module_access(&name, label, &location, access_location);
 
@@ -1106,15 +1169,9 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 .environment
                 .imported_modules
                 .get(module_alias)
-                .ok_or_else(|| Error::UnknownModule {
-                    name: module_alias.to_string(),
-                    location: *module_location,
-                    known_modules: self
-                        .environment
-                        .importable_modules
-                        .keys()
-                        .map(|t| t.to_string())
-                        .collect(),
+                .ok_or_else(|| {
+                    self.environment
+                        .err_unknown_module(module_alias.to_string(), *module_location)
                 })?;
 
             let constructor =
@@ -1163,15 +1220,66 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
     }
 
     #[allow(clippy::result_large_err)]
+    fn infer_type_constructor_access(
+        &mut self,
+        (type_name, type_location): (&str, Span),
+        (label, label_location): (&str, Span),
+    ) -> Result<TypedExpr, Error> {
+        self.environment.increment_usage(type_name);
+
+        let parent_type = self
+            .environment
+            .module_types
+            .get(type_name)
+            .ok_or_else(|| Error::UnknownType {
+                location: type_location,
+                name: type_name.to_string(),
+                types: self.environment.known_type_names(),
+            })?;
+
+        let module_name = parent_type.module.clone();
+
+        self.infer_inner_type_constructor_access(
+            (module_name.as_str(), type_location),
+            (type_name, type_location),
+            (label, label_location),
+        )
+    }
+
+    #[allow(clippy::result_large_err)]
+    fn infer_inner_type_constructor_access(
+        &mut self,
+        (module_name, module_location): (&str, Span),
+        (type_name, type_location): (&str, Span),
+        (label, label_location): (&str, Span),
+    ) -> Result<TypedExpr, Error> {
+        self.environment.get_fully_qualified_value_constructor(
+            (module_name, module_location),
+            (type_name, type_location),
+            (label, label_location),
+        )?;
+
+        self.environment.unused_modules.remove(module_name);
+
+        self.infer_module_access(
+            &self
+                .environment
+                .local_module_name(module_name, module_location)?,
+            label.to_string(),
+            &type_location,
+            label_location,
+        )
+    }
+
+    #[allow(clippy::result_large_err)]
     fn infer_record_access(
         &mut self,
         record: UntypedExpr,
         label: String,
         location: Span,
     ) -> Result<TypedExpr, Error> {
-        // Infer the type of the (presumed) record
+        // Infer the type of the (presumed) record.
         let record = self.infer(record)?;
-
         self.infer_known_record_access(record, label, location)
     }
 
@@ -2563,15 +2671,9 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                     .environment
                     .imported_modules
                     .get(module_name)
-                    .ok_or_else(|| Error::UnknownModule {
-                        location: *location,
-                        name: module_name.to_string(),
-                        known_modules: self
-                            .environment
-                            .importable_modules
-                            .keys()
-                            .map(|t| t.to_string())
-                            .collect(),
+                    .ok_or_else(|| {
+                        self.environment
+                            .err_unknown_module(module_name.to_string(), *location)
                     })?;
 
                 module